Allow dynamic external interface
When defining a dynamic method name with the
`include Procto.call(:some_method_name)`, only the internal interface
can be changed. The external calls will always be `MyProctoClass::call`.
We have a use case where sometimes we want to interact with the returned
object, often to deal with it as a builder. We would like to be to able
use this in a `MyProctClass::build` style. This is to facilitate that.

I didn't know how best to structure the tests for this, so I decided to
put them in a new file. If you can think of a way they work better in
your desired test structure, I'm happy to implement that change.

I changed the rubocop.yml more accurately represent the comment that
precedes it.

The Flog threshold was unusably low. It has been upped.